api/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
torch
torchvision
opencv-python-headless
Pillow
numpy
python-multipart
